backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
google-cloud-vision
python-dotenv
requests
